Add post-functions for social media platfforms
I have added three functions for posting on Twitter, Facebook and LinkedIn in post-generator.js. I have also added a set of buttons in form  of a collapsable menu, that allows the user to choose the plattform on which they want to share the post. The menu is inside the share-button-conainer in post-generator.html